trying to fix a bug

+++ src/gst/pipeline.cpp
@@ -464,10 +464,10 @@ void Pipeline::updateListeners(GstMessage *msg)
 }
 
 
-void Pipeline::seekTo(gint64 pos)
+void Pipeline::seekTo(long long pos)
 {
     if (!gst_element_seek (pipeline_, 1.0, GST_FORMAT_TIME, GST_SEEK_FLAG_FLUSH,
-                GST_SEEK_TYPE_SET, pos,
+                GST_SEEK_TYPE_SET, gint64(pos),
                 GST_SEEK_TYPE_NONE, GST_CLOCK_TIME_NONE)) {
         THROW_ERROR("Seek failed!");
     }
